chore: extract resolveBase module and make yolo a data-driven preset
